Fix warnings all over the code

+++ Saffron/Src/Saffron/Renderer/SceneRenderer.cpp
@@ -86,7 +86,7 @@ void SceneRenderer::Init()
 	// Grid
 	const auto gridShader = Shader::Create("Assets/Shaders/Grid.glsl");
 	sData.GridMaterial = MaterialInstance::Create(Material::Create(gridShader));
-	const float gridScale = 16.025, gridSize = 0.025f;
+	const float gridScale = 16.025f, gridSize = 0.025f;
 	sData.GridMaterial->Set("u_Scale", gridScale);
 	sData.GridMaterial->Set("u_Res", gridSize);
 
@@ -173,7 +173,7 @@ std::pair<Ref<TextureCube>, Ref<TextureCube>> SceneRenderer::CreateEnvironmentMa
 
 	Renderer::Submit([envFiltered, cubemapSize]() {
 		const float deltaRoughness = 1.0f / glm::max(static_cast<float>(envFiltered->GetMipLevelCount() - 1.0f), 1.0f);
-		for ( int level = 1, size = cubemapSize / 2; level < envFiltered->GetMipLevelCount(); level++, size /= 2 ) // <= ?
+		for ( int level = 1, size = cubemapSize / 2; level < static_cast<Int32>(envFiltered->GetMipLevelCount()); level++, size /= 2 ) // <= ?
 		{
 			const GLuint numGroups = glm::max(1, size / 32);
 			glBindImageTexture(0, envFiltered->GetRendererID(), level, GL_TRUE, 0, GL_WRITE_ONLY, GL_RGBA16F);
